refactor(config): 拆分 ConfigManager 为多个单一职责的服务类
将原 ConfigManager 类（2329行/71方法）拆分为多个职责单一的服务类：

1. config-types.ts - 配置类型定义（约280行）
2. core-config-manager.ts - 核心配置文件管理（约300行）
   - 配置文件读取、验证、保存
   - 配置初始化、路径解析
   - 事件通知机制

3. mcp-service-config-manager.ts - MCP服务配置管理（约340行）
   - MCP 端点增删改查
   - MCP 服务配置管理
   - 连接参数管理

4. custom-mcp-tool-manager.ts - 自定义MCP工具管理（约370行）
   - 自定义工具添加、删除、验证
   - 工具配置管理

5. tool-stats-manager.ts - 工具统计管理（约320行）
   - 工具使用统计更新
   - 并发控制锁机制

6. config-manager-facade.ts - 统一外观类（约540行）
   - 组合各服务功能
   - 保持向后兼容的 API

优点：
- 每个类职责单一，易于理解和维护
- 测试代码更容易组织
- 符合单一职责原则
- 保持向后兼容，不影响现有调用方
